project: first commit
Creates basic structure of efu package and a simple import test.

Project on PyPI will be named easyfota-utils but the package and the CLI
commands is named efu.

Makefile has targets to test and to generate egg and wheel
distributions.

The project will be tested using pytest with coverage and pep8
automatically, so there is no need to manually make these checks.

This software will be released under MIT license, so there is as copy of
it.

Change-Id: I00e58d0beb2d78df715531d243a632f7a926e980
Signed-off-by: Pablo Palácios <ppalacios992@gmail.com>