Makefile: Improved test-functional target
test-functional target now installs efu package in the currently Python
environment. This is necessary since we need to have our console
commands on $PATH (setuptools tests our package before it to be
installed so it is not possible to run functional tests bundled to
setuptools).

There is a dumb test just to make CI works for now but it will be removed
as soon as new tests appear.

Change-Id: I52905c8d9aad3e00d8f791d111acbf9d23ed71fe
Signed-off-by: Pablo Palácios <ppalacios992@gmail.com>