SignatureV1: adds _hashed_canonical_request method
This method is necessary since the planned Request class will not be
responsible for it.

Since this modification changes the hardcoded hashes in tests (and since
I could not remember how the hashes in tests were made) the affected
methods were updated (and all hardcoded hashes were documented so it is
easy to understand how they should be made).

Change-Id: I143937d001c9719fc105cf3524405055cb01361d
Signed-off-by: Pablo Palácios <ppalacios992@gmail.com>