SignatureV1: improves final signature and behavior
The timestamp used to calculate the signature was added into the final
signature. In this way, all the elements to reproduce the signature are
present in itself.

Also, SignatureV1 doesn't add the final signature into a request
anymore. Instead, it has a new signature property whose value can be
attached in any place by any other class.

This new behavior can be seen in the Request class, which had to be
updated to follow this change.

Change-Id: Ic9dd8c64dc06c44f21c81e348a80895efd08f095
Signed-off-by: Pablo Palácios <ppalacios992@gmail.com>