httpmock: refactors HTTP request handler
It is now much more robust and simpler to use in tests. Instead of
overwriting directly the RequestHandler attributes, it is now possible
to register many responses as needed using its API.

It is also able to register a history of requests so we can check how
many requests were done and how they were made.

Change-Id: I85151b2bf4576d0d146aaefeef475cf5ea43d5b4
Signed-off-by: Pablo Palácios <ppalacios992@gmail.com>