upload: create first version of upload cli command
This version loads a package file with the list of files that must be
uploaded.

It sends a first request to server asking for urls to upload the
files. It already handles if it should upload a file or not when the
file already exists in server. Also, it is able to send files in
chunks based on server availability.

Finally, it sends a finish request to check if everything was
successfully evaluated on server.

For now, this first version is not able to restart a request if
something went wrong.

Change-Id: Iacbac4b605a5c04d4fbcc880f6f643aec190feac
Signed-off-by: Pablo Palácios <ppalacios992@gmail.com>