cli: replaced cli parser argparse by click package
argparse has a bug (http://bugs.python.org/issue9351) that is not solved
on Python 3.4. This bug makes hard to create cli programs with nested
commands.

click package (http://click.pocoo.org/6/) is developed by the same
people that develops other famous packages (such as flask, jinja and
werkzug). This package is made to create programs with nested commands
easily and in fact, it simplified the complexity of our code and stripped
many lines of it.

Change-Id: Ie1c16a655e7126bd264022fe2ebfd892485c4d20
Signed-off-by: Pablo Palácios <ppalacios992@gmail.com>