tests: creates ServerMocker class
This class was created to decouple server fixture generation from
BaseTransactionTestCase. This is necessary since we also need to test efu
from outside unittest scope.

With this change ServerMocker has all the methods that generates server
fixtures and BaseTransactionTestCase has a new attribute called fixture
that holds an instance of this new created class.

Change-Id: I8689706210d8acf488b56be86d69e09fb3936fa7
Signed-off-by: Pablo Palácios <ppalacios992@gmail.com>