Config: fix efu config command related bugs
efu config set and get command were reviewed since it was not working
properly. Until this patch, it was not possible to update a
configuration entry but this is solved now.

Also, config tests were not made correctly since it was not excluding
the config file generated. To solve this issue in a safer way, the
environment variable approach used with upload server url was
implemented. Since many test cases access efu configuration file, a
mixin was created to set this variable and avoid repetition.

Change-Id: I245b92fe165d95361c96bfa309f4ba8181acb7e1
Signed-off-by: Pablo Palácios <ppalacios992@gmail.com>