File: refactored file validation
A __new__ method was implemented in File to raise an exception if the
passed file does not exists. This eliminate the _validate_file since
validation now occurs before the object instantiation.

The benefit is to have a clearer code and to have certain at object
instantiation time that the file exists.

Change-Id: Ia4863d8d8e0c2e6371162a0b684efb13a256b634
Signed-off-by: Pablo Palácios <ppalacios992@gmail.com>