File: moved chunk number generation to File
FileChunk class is no more responsible for generating the chunk
numbers. This was a problem since the FileChunk._number was a class
attribute which could be messed with parallel File instantiation.

Now, FileChunk expects explicitly a number and each File instance is
responsible for generating and controlling this number.

This modification simplifies codebase (no need for reseting chunk
numbers) and allows parallel File creation.

Change-Id: I9c25e4b2ca3c6afa41291af52e983c68d8565b48
Signed-off-by: Pablo Palácios <ppalacios992@gmail.com>