package: creates explicty and interactive add cmd
This commit creates 'efu add' cli command which works in two modes:
interactive and explicity.

In interactive mode, user just pass the filename as parameter and efu will
prompt for the missing options.

In explicty mode, user must pass all required options for a given
install mode.

This implementation of 'two-mode' command supports parameter dependency,
lazy validation and lazy default values. It was necessary to implement
these lazy features because click package makes a lot of magic which
obfuscate what users really typed at console.

Also, the current implementation does not glue the created image with
final metadata file. The reason for this is the size of complexity
involved to create a two-mode command with click.

Change-Id: I46ad28482359408e8068335c8a64247d2a6d6200
Signed-off-by: Pablo Palácios <ppalacios992@gmail.com>