pull: updated pull command according to specs
'efu pull' now works without passing a product id. For now, it expects a
clean package file to exists in the current working directory.

If a package exists, the operation will abort. Also, as previous
version, if a file to be downloaded exists locally but with a different
content, the operation will abort too.

This improved version also handles better exit codes as also error
messages to user.

Given that a new battery of tests was introduced, PullMockMixin was
created in tests/base.py to aggregate common fixtures between tests.

Change-Id: I311c767ca2c296b1a5196e2885f7899b24bd826a
Signed-off-by: Pablo Palácios <ppalacios992@gmail.com>