Package: implemented array of array of objects
It was necessary to create two new classes to achive this nicely. One
for represent an array of objects (ObjectList) and another one to
represent the array of arrays (ObjectManager). This last one is attached
into Package class and it is responsible for managing package
objects (and this is the reason for its name).

For now, none of efu API clients (cli and repl) supports adding more
than one list per package. This will be implemented in a next commit.

Change-Id: I2ec7734916755ba4ba5dcea8caddcbc9d89fa1c7
Signed-off-by: Pablo Palácios <ppalacios992@gmail.com>