object: added new behaviour for ObjectManager
ObjectManager now has some magic to identify if user is dealing with
single or active-backup mode.

With this new feature, it is not necessary to specify an installation
set index anymore when working in single mode.

In the other way, if user has an active-backup package, our manager will
always raise an exception if index is not specified.

Change-Id: Ib2e367620c596eb62887130fce4fb5c1a2ddaa96
Signed-off-by: Pablo Palácios <ppalacios992@gmail.com>