core: adds basic support for install-if-different
"install-if-different" feature was added partially into efu. There is
already space for it within core: We are already able to import/export
from/to a dot efu file as well from/to a metadata object. CLI and REPL
already have bindings to manage this feature.

efu is not currently able to find versions from objects using this
feature.

Change-Id: If3404e070edc054ec33d6ab5f7b1d8971832559e
Signed-off-by: Pablo Palácios <ppalacios992@gmail.com>