core: removed Product class
Product class was an unecessary code that was in the middle of some
actions making efu to work unexpectedly. With this modification, user
are free to set a new product UID without compromising dot-efu.

Also, product is no longer a requirement to start using efu on cli. Like
in repl, if you don't have anything setted, it will create a new package
for you.

Change-Id: I04dc60e262a2b47592c0c6102474ef5fd0dbee82
Signed-off-by: Pablo Palácios <ppalacios992@gmail.com>