Package: moved get_status to a class method
Since it is not necessary a product UID to request a package status,
we only need the package UID to get package status. Furthermore,
Package.get_status does not have any relation with a Package instance
so we can remove uid from Package.__init__ and promote get_status to
a class method.

Signed-off-by: Pablo Palácios <ppalacios992@gmail.com>