core: removed objects.InstallationSet class
With this modification ObjectsManager now operates on a different data
structure. Instead of using a list of installation sets, we now use a
list of objects where each objects entry has all configurations for
all installations sets.

In other words, we move from:
    [
        [a1, b1],
        [a2, b2],
    ]

To:
    [
        [a1, a2],
        [b1, b2],
    ]

This must be interpreted as a implementation detail since it does not
reflect on dot uhu file. The point here is to operate objects in a
easier and safer way.

Signed-off-by: Pablo Palácios <ppalacios992@gmail.com>